// _ur_addons/webplay/_stash/_types/dataset.d.ts
export type UR_EntID = string;
export type UR_EntID_Obj = { _id: UR_EntID };
/// we use various object conventions
export type DataObj = { [key: string]: any };
export type ErrObj = { error?: string; errorCode?: string; errorInfo?: string };
export type ReturnObj = DataObj | ErrObj;
/// we use UR_DataMethod functions to modify data and datasets
export type UR_DataMethod = (data: DataObj, options?: DataObj) => ReturnObj;
/// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
/// an UR_Item is a union of MatchObj with UR_EntID
export type UR_Item = UR_EntID_Obj & DataObj; // { _id: UR_EntID; [key: string]: any }
export type UR_Doc = UR_Item; // doc is a single item
/// there are multiple ways to organize UR_Items into a "bag"
export type UR_ItemList = UR_Item[];
export type UR_DocFolder = { [_id: UR_EntID]: UR_Doc };
// the document manager has named Documents that are in a
/// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
/// REFS identify a unique "bag" of items in a data model, despite the type
/// of bag it is (e.g. documents, itemlists, etc.)
export type UR_BagRef = string;